Support ARM Ubuntu

+++ features/chrome.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 
 # install chrome
-curl -L https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb -o /tmp/chrome.deb
-apt-get install --yes /tmp/chrome.deb
-rm /tmp/chrome.deb
+if [[ "$(dpkg --print-architecture)" = arm* ]]; then
+  apt-get install --yes chromium-browser
+else
+  curl -L https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb -o /tmp/chrome.deb
+  apt-get install --yes /tmp/chrome.deb
+  rm /tmp/chrome.deb
+fi
